requirements.txt
grpcio
grpcio-tools
numpy==1.24.2
protobuf==4.22.0
PyYAML==6.0
torch==1.13.1
pytorch-crf
omegaconf==2.3.0
transformers==4.26.1